requirements.txt
attr==0.3.2
beautifulsoup4==4.13.3
ConfigParser==7.1.0
contextlib2==21.6.0
cryptography==44.0.0
docutils==0.21.2
fastapi==0.115.8
filelock==3.17.0
HTMLParser==0.0.2
httpx==0.28.1
importlib_metadata==8.0.0
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.5
keyring==25.6.0
mock==5.1.0
newspaper3k==0.2.8
openai==1.61.1
Pillow==11.1.0
pinecone==5.4.2
protobuf==5.29.3
pydantic==2.10.6
pyOpenSSL==25.0.0
pyOpenSSL==25.0.0
python-dotenv==1.0.1
railroad==0.5.0
redis==5.2.1
sentence_transformers==3.4.1
Sphinx==8.1.3
SQLAlchemy==2.0.32
starlette==0.45.3
thread==2.0.5
tornado==6.4.2
urllib3_secure_extra==0.1.0
uvicorn==0.34.0

